refactor: rename `@reflect/dynamic-modules` to `@reflect/modules`
