+++ packages/core/src/ai/audio-memo-format.ts
@@ -94,7 +94,7 @@ export async function formatAudioMemoTranscript(
   }
 
   try {
-    const { generateText, Output } = await import('@reflect/dynamic-modules/ai')
+    const { generateText, Output } = await import('@reflect/modules/ai')
     const result = await generateText({
       model: await languageModel(config, request.credentials.apiKey, request.fetchFn ?? fetch),
       output: Output.object({ schema: formattedAudioMemoSchema }),
