+++ packages/core/src/ai/chat/tools.ts
@@ -1,4 +1,4 @@
-import type { Tool, TypedToolCall, TypedToolResult } from '@reflect/dynamic-modules/ai'
+import type { Tool, TypedToolCall, TypedToolResult } from '@reflect/modules/ai'
 import { z } from 'zod'
 import { readNote } from '../../graph/commands'
 import { retrieve, type RetrievalHit, type RetrieveOptions } from '../../embeddings/retrieve'
