+++ packages/core/src/ai/describe-asset.test.ts
@@ -1,5 +1,5 @@
-import { APICallError } from '@reflect/dynamic-modules/ai'
-import { MockLanguageModelV3 } from '@reflect/dynamic-modules/ai/test'
+import { APICallError } from '@reflect/modules/ai'
+import { MockLanguageModelV3 } from '@reflect/modules/ai/test'
 import type { LanguageModelV3CallOptions, LanguageModelV3Usage } from '@ai-sdk/provider'
 import { describe, expect, it, vi } from 'vitest'
 import type { AiProviderConfig } from '../settings/schema'
