+++ packages/core/src/ai/describe-page.ts
@@ -1,4 +1,4 @@
-import type { UserContent } from '@reflect/dynamic-modules/ai'
+import type { UserContent } from '@reflect/modules/ai'
 import { z } from 'zod'
 import { ReflectError } from '../errors'
 import { wikiLinkSafe } from '../markdown/edit'
@@ -83,10 +83,7 @@ export function isDescriptionRejected(value: unknown): value is DescriptionRejec
 
 function classify(
   cause: unknown,
-  sdk: Pick<
-    typeof import('@reflect/dynamic-modules/ai'),
-    'APICallError' | 'NoObjectGeneratedError'
-  >,
+  sdk: Pick<typeof import('@reflect/modules/ai'), 'APICallError' | 'NoObjectGeneratedError'>,
 ): Error {
   if (sdk.APICallError.isInstance(cause)) {
     const status = cause.statusCode ?? 0
@@ -155,7 +152,7 @@ export function normalizedPageTitle(candidate: string): string | null {
  * itself.
  */
 export async function describePage(request: DescribePageRequest): Promise<PageEnrichment> {
-  const sdk = await import('@reflect/dynamic-modules/ai')
+  const sdk = await import('@reflect/modules/ai')
   const content: UserContent = [{ type: 'text', text: describePrompt(request) }]
   if (request.screenshotBase64) {
     content.push({
