+++ packages/core/src/ai/language-model.ts
@@ -1,4 +1,4 @@
-import type { LanguageModel } from '@reflect/dynamic-modules/ai'
+import type { LanguageModel } from '@reflect/modules/ai'
 import type { AiProviderConfig } from '../settings/schema'
 import { anthropicDirectBrowserAccessHeaders } from './anthropic-headers'
 import { APP_REVIEW_STUB_KEY, createDemoModel } from './app-review-demo'
@@ -23,11 +23,11 @@ export async function languageModel(
   }
   switch (config.provider) {
     case 'openai': {
-      const { createOpenAI } = await import('@reflect/dynamic-modules/ai-sdk/openai')
+      const { createOpenAI } = await import('@reflect/modules/ai-sdk/openai')
       return createOpenAI({ apiKey, fetch: fetchFn })(config.model)
     }
     case 'anthropic': {
-      const { createAnthropic } = await import('@reflect/dynamic-modules/ai-sdk/anthropic')
+      const { createAnthropic } = await import('@reflect/modules/ai-sdk/anthropic')
       return createAnthropic({
         apiKey,
         fetch: fetchFn,
@@ -35,11 +35,11 @@ export async function languageModel(
       })(config.model)
     }
     case 'google': {
-      const { createGoogle } = await import('@reflect/dynamic-modules/ai-sdk/google')
+      const { createGoogle } = await import('@reflect/modules/ai-sdk/google')
       return createGoogle({ apiKey, fetch: fetchFn })(config.model)
     }
     case 'openrouter': {
-      const { createOpenAI } = await import('@reflect/dynamic-modules/ai-sdk/openai')
+      const { createOpenAI } = await import('@reflect/modules/ai-sdk/openai')
       return createOpenAI({
         apiKey,
         fetch: fetchFn,
@@ -49,8 +49,7 @@ export async function languageModel(
       }).chat(config.model)
     }
     case 'openai-compatible': {
-      const { createOpenAICompatible } =
-        await import('@reflect/dynamic-modules/ai-sdk/openai-compatible')
+      const { createOpenAICompatible } = await import('@reflect/modules/ai-sdk/openai-compatible')
       return createOpenAICompatible({
         name: OPENAI_COMPATIBLE_PROVIDER_ID,
         baseURL: config.baseUrl,
