+++ packages/modules/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@reflect/dynamic-modules",
+  "name": "@reflect/modules",
   "version": "0.1.0",
   "private": true,
   "type": "module",
