+++ packages/modules/src/ai-sdk/openai-compatible.ts
