+++ packages/modules/src/ai.ts
