+++ tsconfig.json
@@ -14,7 +14,7 @@
       "path": "./packages/db/tsconfig.json"
     },
     {
-      "path": "./packages/dynamic-modules/tsconfig.json"
+      "path": "./packages/modules/tsconfig.json"
     },
     {
       "path": "./packages/utils/tsconfig.json"
